+++ src/main/java/no/vegvesen/nvdbapi/client/model/roadobjects/attribute/IntegerEnumAttribute.java
@@ -1,0 +1,7 @@
+package no.vegvesen.nvdbapi.client.model.roadobjects.attribute;
+
+public class IntegerEnumAttribute extends EnumAttribute<Integer> {
+    public IntegerEnumAttribute(int id, Integer enumId, Integer value) {
+        super(id, enumId, value);
+    }
+}