+++ src/main/java/no/vegvesen/nvdbapi/client/model/roadobjects/attribute/RealEnumAttribute.java
@@ -1,0 +1,7 @@
+package no.vegvesen.nvdbapi.client.model.roadobjects.attribute;
+
+public class RealEnumAttribute extends EnumAttribute<Double> {
+    public RealEnumAttribute(int id, Integer enumId, Double value) {
+        super(id, enumId, value);
+    }
+}