+++ src/main/java/no/vegvesen/nvdbapi/client/model/roadobjects/attribute/StringEnumAttribute.java
@@ -1,0 +1,7 @@
+package no.vegvesen.nvdbapi.client.model.roadobjects.attribute;
+
+public class StringEnumAttribute extends EnumAttribute<String> {
+    public StringEnumAttribute(int id, Integer enumId, String value) {
+        super(id, enumId, value);
+    }
+}